% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_grp_s3.R
\name{as.var_grp_df}
\alias{as.var_grp_df}
\title{The \code{var_grp_df} dataframe subtype}
\usage{
as.var_grp_df(df, z, y, x)
}
\arguments{
\item{df}{a dataframe}

\item{z}{the z columns (e.g. cohort) as a list of columns}

\item{y}{the y columns (e.g. response) as a list of columns}

\item{x}{the x columns (e.g. predictor) as a list of columns}
}
\description{
This is like a grouped data frame but with 3 grouping dimensions. These are
labelled z, y, and x and relate to as \code{z} (i.e. group, or cohort), \code{y} (i.e.
subgroup, or response) and \code{x} (i.e. data). In some configurations,
only \code{z} and \code{x} are non-empty. The purpose of this is to make some group / subgroup
data operations consistent. An example is running multiple models across
different bootstraps from example.
}
\examples{
tmp = as.var_grp_df(iris, 
  c("Species"), 
  c("Sepal.Width", "Sepal.Length"), 
  c("Petal.Width", "Petal.Length"))
# print.var_grp_df(tmp)
glimpse.var_grp_df(tmp)

}
\concept{var_group}
