% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consistent.R
\name{check_consistent}
\alias{check_consistent}
\title{Check function parameters are conform to a set of rules}
\usage{
check_consistent(..., .env = rlang::caller_env())
}
\arguments{
\item{...}{a set of rules to check either as \code{x=y+z}, or \code{x>y}. Single \code{=}
assignment is checked for equality using \code{identical} otherwise the
expressions are evaluated and checked they all are true. This for
consistency with \code{resolve_missing} which only uses assignment, and ignores
logical expressions.}

\item{.env}{the environment to check in}
}
\value{
nothing, throws an informative error if the checks fail.
}
\description{
Check function parameters are conform to a set of rules
}
\examples{
testfn = function(pos=NULL, neg=NULL, n=NULL) {
  check_consistent(pos=n-neg, neg=n-pos, n=pos+neg, n>pos, n>neg)
}

testfn(pos = 1:4, neg=4:1, n=rep(5,4))
try(testfn(pos = 1:4, neg=5:2, n=rep(5,4)))
}
