% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_check_date.R
\name{check_date}
\alias{check_date}
\title{Checks a set of variables can be coerced to a date and coerces them}
\usage{
check_date(
  ...,
  .message = "`{param}` is not a date: ({err}).",
  .env = rlang::caller_env()
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:as.Date]{base::as.Date}}
  \describe{
    \item{\code{x}}{an object to be converted.}
  }}

\item{.message}{a glue spec containing \code{{param}} as the name of the parameter and \code{{err}} the cause fo the error}

\item{.env}{the environment to check (defaults to calling environment)}
}
\value{
nothing. called for side effects. throws error if not all variables can be coerced.
}
\description{
Checks a set of variables can be coerced to a date and coerces them
}
\examples{
a = c(1:4L)
b = c("1",NA,"3.3")
f = NULL
g = NA
check_numeric(a,b,f,g)

c = c("dfsfs")
try(check_numeric(c,d, mean))
}
\concept{params_check}
