\name{class_name}
\alias{class_name}
\alias{class_name.class_desc}
\alias{class_name.classes_desc}
\title{Find class names}
\usage{
  class_name(x)

  \method{class_name}{class_desc} (x)

  \method{class_name}{classes_desc} (x)
}
\arguments{
  \item{x}{Object of class \sQuote{class_desc} or
  \sQuote{classes_desc}.}
}
\value{
  Character vector.
}
\description{
  Find the name(s) of the documented class(es) within a
  \sQuote{class_desc} or \sQuote{classes_desc} object.
}
\keyword{internal}

