% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_syms.R
\name{col_syms}
\alias{col_syms}
\title{Column names as symbols}
\usage{
col_syms(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a list of symbols
}
\description{
Column names as symbols
}
\examples{
intersect(colsyms(iris), ensyms2(tidyselect::starts_with("S"), .df=iris))
}
\concept{var_group}
