% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_formals.R
\name{doc_formals}
\alias{doc_formals}
\title{Document default formal parameters in a function}
\usage{
doc_formals(fn, param, one_of = TRUE)
}
\arguments{
\item{fn}{a function in the current package}

\item{param}{a parameter (usually the same as the param block)}

\item{one_of}{decorate the string with a "default"/"one of" depending on type}
}
\value{
a formatted string for inclusion in a Roxygen block
}
\description{
this is intended to be used in a Roxygen tag to document the
default value or allowable values of a function
}
\examples{
test_fn = function(
    arg = c("option one","option two","option three"),
    def = 123,
    geh = def*2
) {
  arg = match.arg(arg)
}

doc_formals(test_fn, arg)
doc_formals(test_fn, def)
doc_formals(test_fn, geh)
}
