% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escalate.R
\name{escalate}
\alias{escalate}
\title{Cause warnings to create an error}
\usage{
escalate(expr)
}
\arguments{
\item{expr}{expression to evaluate}
}
\value{
the evaluated expression or an error
}
\description{
The opposite of \code{suppressWarnings()}. This will immediately error if a warning
if thrown by \code{expr}. This is useful to track down the source of a random
and to prevent Rs permissive approach to data transformations. It is also
useful to identify where in the code a intermittent \code{rlang} warning is being
issued once every 8 hours.
}
\examples{
try(escalate(as.integer("ASDAS")))

try(escalate(rlang::warn("test", .frequency="regularly", .frequency_id = "asdasdasasdd")))
try(escalate(rlang::warn("test", .frequency="regularly", .frequency_id = "asdasdasasdd")))
try(escalate(rlang::warn("test", .frequency="regularly", .frequency_id = "asdasdasasdd")))
try(escalate(rlang::warn("test", .frequency="regularly", .frequency_id = "asdasdasasdd")))

# options("rlib_warning_verbosity"=NULL)
# options("rlib_warning_verbosity"="verbose")
# "lifecycle_verbosity"="warning"
}
