% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escalateWarnings.R
\name{escalateWarnings}
\alias{escalateWarnings}
\title{Cause warnings to create an error}
\usage{
escalateWarnings(expr)
}
\arguments{
\item{expr}{expression to evaluate}
}
\value{
the evaluated expression of an error
}
\description{
Cause warnings to create an error
}
\examples{
try(escalateWarnings(as.integer("ASDAS")))
}
