% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findRefs}
\alias{findRefs}
\title{find reference of function within a package}
\usage{
findRefs(pkg, fn)
}
\description{
在指定包中查找调用 fn 的函数
}
\examples{
findRefs('knitr', 'auto_out_name')
}
\seealso{
Other package: \code{\link{install_packages}};
  \code{\link{md2context}}; \code{\link{md_roclet}};
  \code{\link{pkgBackup}}; \code{\link{pkgLoad}};
  \code{\link{pkgRestore}}; \code{\link{roxygen_fix}};
  \code{\link{viewCode}}
}

