% Generated by roxygen2 (4.0.1): do not edit by hand
\name{md2context}
\alias{md2context}
\title{将 `markdown` 文档转换为`ConTeXt`}
\usage{
md2context(fn, template = "r_pkg", toPdf = TRUE)
}
\arguments{
\item{fn}{待转换的文件名}

\item{template}{`ConTeXt` 模板，默认为 `r_pkg`}

\item{toPdf}{是否自动调用 `ConTeXt` 编译 pdf。默认为 TRUE。}
}
\description{
调用 `pandoc` 将 `markdown` 文档转换为 `ConTeXt` 格式。
}
\note{
`pandoc` 及 `ConTeXt` 已在系统路径中。`ConTeXt` 环境初始化命令为 `setuptex2.bat`。
}
\seealso{
oxygenize_markdown

Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md_roclet}};
  \code{\link{pkgBackup}}; \code{\link{pkgLoad}};
  \code{\link{pkgRestore}}; \code{\link{roxygen_fix}};
  \code{\link{viewCode}}
}

