% Generated by roxygen2 (4.0.1): do not edit by hand
\name{md_roclet}
\alias{md_roclet}
\title{oxygenize `markdown` 格式文档}
\usage{
md_roclet(pkgdir)
}
\arguments{
\item{pkgdir}{需要生成文档的包路径}
}
\description{
根据 `R` 脚本中的 `oxygen` 语法生成 `markdown` 格式说明文档。
代码重复利用自 `roxygen2` 包。
}
\details{
返回值为生成的 `markdown` 文件名
}
\examples{
fn = md_roclet(base_path)
md2context(fn)
}
\seealso{
Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md2context}};
  \code{\link{pkgBackup}}; \code{\link{pkgLoad}};
  \code{\link{pkgRestore}}; \code{\link{roxygen_fix}};
  \code{\link{viewCode}}
}

