% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_fn.R
\name{optional_fn}
\alias{optional_fn}
\title{Get an optional function without triggering a CRAN warning}
\usage{
optional_fn(
  pkg,
  name,
  alt = function(...) {
     stop("function `", pkg, "::", name, "(...)` not available")

    }
)
}
\arguments{
\item{pkg}{the package name (or the function name as \code{"pkg::fn"})}

\item{name}{the function you wish to use (if not specified in \code{pkg})}

\item{alt}{an alternative function that can be used if the requested one is
not available. The default throws an error if the package is not available,
but a fallback can be used instead.}
}
\value{
the function you want if available or the alternative
}
\description{
You want to use a function if it is installed but don't want it to
be installed as part of your package and you don't want to reference it
as part of the Imports or Suggests fields in a package DESCRIPTION.
}
\examples{
# use openSSL if installed:
fn = optional_fn("openssl", "md5", alt = digest::digest)
as.character(fn(as.raw(c(1,2,3))))

fn2 = optional_fn("asdasdadsda::asdasdasd")
try({
  # block to execute if asdasdadsda::asdasdasd() is available
  fn2()
}) 
# add silent=TRUE to hide error message

# this function does not exists and so the alternative is used instead.
fn3 = optional_fn("asdasdadsda::asdasdasd", ~ message("formula alternative"))
fn3()

}
