% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgBackup}
\alias{pkgBackup}
\title{备份已安装包名}
\usage{
pkgBackup()
}
\description{
将已安装的包名备份并保存到当前路径下的 `Rpkgs.RData` 文件。
}
\seealso{
Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md2context}};
  \code{\link{md_roclet}}; \code{\link{pkgLoad}};
  \code{\link{pkgRestore}}; \code{\link{roxygen_fix}};
  \code{\link{viewCode}}
}

