% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgLoad}
\alias{pkgLoad}
\title{载入或安装包}
\usage{
pkgLoad(pkg)
}
\description{
载入或安装包
}
\seealso{
Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md2context}};
  \code{\link{md_roclet}}; \code{\link{pkgBackup}};
  \code{\link{pkgRestore}}; \code{\link{roxygen_fix}};
  \code{\link{viewCode}}
}

