% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkgRestore}
\alias{pkgRestore}
\title{根据已备份的包名称，安装包}
\usage{
pkgRestore()
}
\description{
根据已备份的包名称，安装包
}
\seealso{
Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md2context}};
  \code{\link{md_roclet}}; \code{\link{pkgBackup}};
  \code{\link{pkgLoad}}; \code{\link{roxygen_fix}};
  \code{\link{viewCode}}
}

