% Generated by roxygen2 (4.0.1): do not edit by hand
\name{roxygen_fix}
\alias{roxygen_fix}
\title{修正 `Rd` 文件的中文乱码}
\usage{
roxygen_fix(pkgdir)
}
\arguments{
\item{pkgdir}{需要修正编码的包路径，或 \code{package_path/man/}。}
}
\description{
默认 \code{roxygen2} 生成的文件为 ASCII 编码。该函数将 \code{Rd}  转换为 \code{UTF-8}，
并在首行添加 \code{\\encoding}。
}
\details{
\code{roxygen2::roxygenize} 写入的 `rd` 文件编码取决于 `R` 文件的编码。若 `R` 文件为 `UTF-8` 编码，则
`rd` 文件也为 `UTF-8` 编码。此时，无需再做编码转换，在 `DESCRIPTION` 文件中添加 `Encoding`
字段或直接为 `rd` 文件添加 `\\encoding{}` 即可。对于这种情形，无需调用本函数。

\code{roxygen2} 生成 `R` help 支持中文说明

+ 方式一：设置 `R` 文件及 `DESCRIPTION` 为 \code{utf-8} 编码。

+ 方式二：

- \code{rd} 文件头添加 \code{\\encoding{utf-8}}

- \code{rd} 文件编码为 \code{utf-8}，无 BOM

- \code{title} 的中文在 R help index 中仍为乱码，通过替换 \code{tools::.writePkgIndices()} 解决。

- 若 `DESCRIPTION` 文件中设定了 \code{Encoding:utf-8}，则 \code{rd} 文件无需添加 \code{\\encoding{utf-8}}
}
\seealso{
Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md2context}};
  \code{\link{md_roclet}}; \code{\link{pkgBackup}};
  \code{\link{pkgLoad}}; \code{\link{pkgRestore}};
  \code{\link{viewCode}}
}

