% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_group_count.R
\name{var_group_count}
\alias{var_group_count}
\title{The number of major groups (\code{z} categories) in a \code{var_grp_df}}
\usage{
var_group_count(var_grp_df)
}
\arguments{
\item{var_grp_df}{the var_grp dataframe}
}
\value{
a count of groups
}
\description{
The number of major groups (\code{z} categories) in a \code{var_grp_df}
}
\examples{
tmp = iris \%>\% dplyr::group_by(Species) \%>\% var_group(. ~ Petal.Width + Sepal.Width)
tmp \%>\% var_group_count()
}
\concept{var_group}
