% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_grp_s3.R
\name{var_group_formula}
\alias{var_group_formula}
\title{Export \code{var_group} metadata as a formula}
\usage{
var_group_formula(var_grp_df)
}
\arguments{
\item{var_grp_df}{a \code{var_group} dataframe}
}
\value{
a formula like \code{y1 + y2 ~ x1 + x2 + ...}
}
\description{
Produces the \code{y} and \code{x} terms of a \code{var_grp_df} as a formula for
potentially using in a model or another \code{var_group}
}
\concept{var_group}
