% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_grp_s3.R
\name{var_grp_s3}
\alias{var_grp_s3}
\alias{glimpse.var_grp_df}
\alias{format.var_grp_df}
\alias{print.var_grp_df}
\alias{is.var_grp_df}
\title{\code{var_grp_df} S3 Methods}
\usage{
glimpse.var_grp_df(x, ...)

\method{format}{var_grp_df}(x, ...)

\method{print}{var_grp_df}(x, ...)

is.var_grp_df(x, ...)
}
\arguments{
\item{x}{a \code{var_grp_df} dataframe}

\item{...}{passed to generic functions}
}
\description{
\code{var_grp_df} S3 Methods
}
\section{Functions}{
\itemize{
\item \code{glimpse.var_grp_df()}: glimpse

\item \code{format(var_grp_df)}: format

\item \code{print(var_grp_df)}: print

\item \code{is.var_grp_df()}: is

}}
