% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_grps.R
\name{var_grps}
\alias{var_grps}
\title{Extract grouping info frm a \code{var_grp_df}}
\usage{
var_grps(var_grp_df)
}
\arguments{
\item{var_grp_df}{the dataframe}
}
\value{
a list of lists containing the \code{x},\code{y}, and \code{z} column sets as symbol lists
}
\description{
Extract grouping info frm a \code{var_grp_df}
}
\concept{var_group}
