% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_has_groups.R
\name{var_has_groups}
\alias{var_has_groups}
\title{Does this \code{var_grp_df} have more than one major group?}
\usage{
var_has_groups(var_grp_df)
}
\arguments{
\item{var_grp_df}{a \code{var_group} dataframe}
}
\value{
boolean
}
\description{
Does this \code{var_grp_df} have more than one major group?
}
\concept{var_group}
