% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_group_nest.R
\name{var_subgroup_nest}
\alias{var_subgroup_nest}
\title{Nest a \code{var_grp_df} by the \code{z} and \code{y} columns}
\usage{
var_subgroup_nest(var_grp_df, .key = "data")
}
\arguments{
\item{var_grp_df}{the var_grp dataframe}

\item{.key}{The name of the resulting nested column. Only applicable when
\code{...} isn't specified, i.e. in the case of \code{df \%>\% nest(.by = x)}.

If \code{NULL}, then \code{"data"} will be used by default.}
}
\value{
a nested dataframe with \code{z} and \code{y} columns and a \code{.key} column with
the \code{x} columns nested in it
}
\description{
Nest a \code{var_grp_df} by the \code{z} and \code{y} columns
}
\examples{
tmp = iris \%>\% dplyr::group_by(Species) \%>\% var_group(. ~ Petal.Width + Sepal.Width)
tmp2 = tmp \%>\% var_group_nest()
}
\concept{var_group}
