% Generated by roxygen2 (4.0.1): do not edit by hand
\name{viewCode}
\alias{viewCode}
\title{View code of specified package}
\usage{
viewCode(fun, package)
}
\description{
View code of specified package
}
\seealso{
Other package: \code{\link{findRefs}};
  \code{\link{install_packages}}; \code{\link{md2context}};
  \code{\link{md_roclet}}; \code{\link{pkgBackup}};
  \code{\link{pkgLoad}}; \code{\link{pkgRestore}};
  \code{\link{roxygen_fix}}
}

