\docType{data}
\name{xyplot.docu}
\alias{xyplot.docu}
\title{Example RD dataset: xyplot() documentation}
\format{\sQuote{Rd} object as produced by \code{parse_Rd} from the
  \pkg{tools} package.}
\description{
  This \sQuote{Rd} object contains the description in
  \sQuote{Rd} format of the function \code{xyplot} from the
  \pkg{lattice} package, version 0.20-10.
}
\examples{
data(xyplot.docu)
stopifnot(inherits(xyplot.docu, "Rd"), is.list(xyplot.docu))
}
\keyword{datasets}
\keyword{datasets}

